Issue #4: Access controller and action info from interceptor.
Rename InterceptorChain for InterceptorExecution.
Move the Interceptor and InterceptorExecution to the org.jogger.interceptor package.
Create a Controller and Action interfaces in the org.jogger.interceptor package.
Add a getController() and getAction() methods to the InterceptorExecution interface.
Implement the getController() and getAction() methods in the AbstractJoggerServlet.ControllerExecutor.
Add tests to validate new functionality.